backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
pydantic>=2.9.0,<3.0.0
pydantic-settings>=2.5.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
requests>=2.32.0,<3.0.0
numpy>=2.1.0,<3.0.0
loguru>=0.7.2,<1.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib>=1.7.4,<2.0.0
python-multipart>=0.0.9,<1.0.0
neo4j>=5.25.0,<6.0.0
langchain>=0.3.0,<0.4.0
langchain-openai>=0.2.0,<0.3.0
faiss-cpu>=1.9.0 ; platform_system != "Windows"
httpx>=0.27.0,<1.0.0
sqlalchemy>=2.0.36,<3.0.0
psycopg[binary]>=3.2.3,<4.0.0
redis>=5.2.0,<6.0.0
celery>=5.4.0,<6.0.0
boto3>=1.35.0,<2.0.0
pypdf>=5.1.0,<6.0.0
python-docx>=1.1.2,<2.0.0
openpyxl>=3.1.5,<4.0.0
PyYAML>=6.0.2,<7.0.0
pytest>=8.3.0,<9.0.0
pytest-asyncio>=0.24.0,<1.0.0

